// fitmateapp/src/api-generated/models/UserSummaryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Simplified user information for search results.
 */
export type UserSummaryDto = {
    /**
     * The user's unique username.
     */
    userName?: string | null;
    /**
     * The user's full name.
     */
    fullName?: string | null;
};